[refactor] #56 client 패키지 내 카카오 api 클래스 위치
