+++ src/main/java/org/lostnomore/backend/auth/oauth/kakao/client/KakaoApiClient.java
@@ -1,4 +1,4 @@
-package org.lostnomore.backend.auth.oauth.kakao;
+package org.lostnomore.backend.auth.oauth.kakao.client;
 
 import org.lostnomore.backend.auth.oauth.kakao.dto.KakaoUserResponse;
 import org.springframework.cloud.openfeign.FeignClient;
